Use the material-design-icons directly from Google's git repository
Google has a version of the material icons specifically intended
for self-hosting at
https://github.com/google/material-design-icons/tree/master/iconfont,
so use these files instead.  The CSS file is, however, modified to
remove the .eot font since this is only used by IE <9, which is
unsupported anyway.  This also adds .woff and .woff2 fonts which
are much smaller and save bandwidth when using modern browsers.